Remove Hiding of UnsupportedCriterionException
The UnsupportedCriterionException should not be hided because the
criteria are an external input. It's always possible that criteria can't
be translated into a FHIR search query string.

Also make SearchQueryStringBuilder testable by replace the side-effected
retrieval of the current time by a clock that can be set to a fixed
clock in tests.

Also remove the integration test, because FHIR query strings can be
tested in the unit test.